Split RPC handles from supporting logic
